docs/gus_projection/requirements.txt
viam-sdk
opencv-python
numpy
Pillow